requirements.txt
# requirements.txt

fastapi
uvicorn
pandas
numpy
matplotlib
seaborn
scikit-learn
python-multipart
openpyxl
xgboost
joblib
imbalanced-learn
python-dotenv